+++ static/splashes/install.php
@@ -54,24 +54,17 @@ wp_cache_flush();
     <h3><?php printf( __( 'WP-Property %s important changes', ud_get_wp_property()->domain ), ud_get_wp_property( 'version' ) ); ?>:</h3>
 
     <ul>
-      <li>Added compatibility with Page Builder by SiteOrigin plugin.</li>
-      <li>Added ability to set single or page templates of current theme for rendering Single Property page instead of predefined property.php.</li>
-      <li>Added ability to disable WP-Property Widget Sidebars.</li>
-      <li>Added Dutch ( Netherlands ) localization.</li>
-      <li>Added <i>Property Overview</i> widget based on <code>[property_overview]</code> shortcode.</li>
-      <li>Added <i>Property Attributes</i> widget which renders the list of property attributes.</li>
-      <li>Added <i>Property Map</i> widget based on <code>[property_map]</code> shortcode.</li>
-      <li>Added <i>List Attachments</i> widget based on <code>[list_attachments]</code> shortcode.</li>
-      <li>Added <code>[property_meta]</code> shortcode and Property Meta widget.</li>
-      <li>Added <code>[property_terms]</code> shortcode and Property Terms widget.</li>
-      <li>Added 'Sort By' and 'Sort Order' options for <i>Child Properties</i> widget.</li>
-      <li>Extended functionality of <code>[property_attributes]</code> shortcode.</li>
-      <li>Updated plugin initialisation logic.</li>
-      <li>Refactored widgets structure and initialisation.</li>
-      <li>Refactored shortcodes structure and initialisation.</li>
-      <li><i>Child Properties</i>, <i>Featured Properties</i>, <i>Latest Properties</i> and <i>Other Properties</i> widgets were deprecated and disabled on new plugin installation. But they still can be activated via Settings.</li>
-      <li>Fixed the bug which prevented to update Add-ons via inline updater on Plugins page.</li>
-      <li>Fixed address validation by provided Google Maps API key.</li>
+      <li>Added function get_property_type() which returns label of property type for current or particular property.</li>
+      <li>Added automatic object cache and plugin cache flushing on plugin's settings updates.</li>
+      <li>Added automatic object cache flushing on property updates.</li>
+      <li>Added automatic plugin cache flushing on plugin activation.</li>
+      <li>Updated localisation files.</li>
+      <li>Updated Russian localisation.</li>
+      <li>Refactored get_property function.</li>
+      <li>Fixed 'Property/Properties' white labels.</li>
+      <li>Fixed plugin's upgrade process.</li>
+      <li>Fixed localisation of javascript files on multi site.</li>
+      <li>Fixed rewrite rules breaking on plugin activation which has caused 404 errors on single property pages.</li>
     </ul>
 
   </div>