+++ changes.md
@@ -8,6 +8,7 @@
 * Refactored get_property function.
 * Fixed 'Property/Properties' white labels.
 * Fixed plugin's upgrade process.
+* Fixed localisation of javascript files on multi site.
 * Fixed rewrite rules breaking on plugin activation which has caused 404 errors on single property pages.
 
 ### 2.1.0 ( October 8, 2015 )